refactor: split fynance/models into one-class-per-file
- neural_network.py → _base.py (BaseNeuralNet) + mlp.py (MultiLayerPerceptron)
- recurrent_neural_network.py → rnn.py, gru.py, lstm.py + _recurrent_base.py
- Rename internals: _RecurrentNeuralNetwork → _RecurrentBase,
  _ForwardLayer → _OutputLayerMixin, _GatedRecurrentUnit → _GRUCell,
  _LongShortTermMemory → _LSTMCell
- Update all internal imports (rolling.py, __init__.py, tests)
- Add test_rnn.py (6 tests for RecurrentNeuralNetwork, previously uncovered)
- Expand GRU and LSTM test suites (train_on, predict_no_grad, int_constructor)
- Update Sphinx RST pages to new module paths; Sphinx build clean
