+++ doc/source/models.recurrent_neural_network.rst
@@ -4,11 +4,32 @@ Recurrent neural networks
 
 Recurrent neural network models (RNN, GRU, LSTM) with walk-forward training support.
 
-.. currentmodule:: fynance.models.recurrent_neural_network
+Vanilla RNN
+===========
+
+.. currentmodule:: fynance.models.rnn
 
 .. autosummary::
    :toctree: generated/
 
    RecurrentNeuralNetwork
+
+Gated Recurrent Unit (GRU)
+==========================
+
+.. currentmodule:: fynance.models.gru
+
+.. autosummary::
+   :toctree: generated/
+
    GatedRecurrentUnit
+
+Long Short-Term Memory (LSTM)
+==============================
+
+.. currentmodule:: fynance.models.lstm
+
+.. autosummary::
+   :toctree: generated/
+
    LongShortTermMemory
