+++ fynance/models/__init__.py
@@ -12,22 +12,29 @@
 
 .. toctree::
 
+    models.mlp
+    models.rnn
+    models.gru
+    models.lstm
     models.attention
     models.econometric_models
-    models.neural_network
-    models.recurrent_neural_network
     models.rolling
 
 """
 
 from . import (
+    _base,
+    _recurrent_base,
     attention,
     econometric_models,
     econometric_models_cy,
-    neural_network,
-    recurrent_neural_network,
+    gru,
+    lstm,
+    mlp,
+    rnn,
     rolling,
 )
+from ._base import BaseNeuralNet
 from .attention import MultiHeadAttention, ScaledDotProductAttention
 from .econometric_models import ARMA, ARMA_GARCH, ARMAX_GARCH, MA, get_parameters
 from .econometric_models_cy import (
@@ -37,12 +44,10 @@ from .econometric_models_cy import (
     MA_cy,
     get_parameters_cy,
 )
-from .neural_network import BaseNeuralNet, MultiLayerPerceptron
-from .recurrent_neural_network import (
-    GatedRecurrentUnit,
-    LongShortTermMemory,
-    RecurrentNeuralNetwork,
-)
+from .gru import GatedRecurrentUnit
+from .lstm import LongShortTermMemory
+from .mlp import MultiLayerPerceptron
+from .rnn import RecurrentNeuralNetwork
 from .rolling import CVResult, RollMultiLayerPerceptron, _RollingBasis
 
 # Frozen public surface for the 1.x series — names listed here are
@@ -65,10 +70,10 @@ __all__ = [
     'ARMAX_GARCH_cy',
     'MA_cy',
     'get_parameters_cy',
-    # neural_network
+    # _base / mlp
     'BaseNeuralNet',
     'MultiLayerPerceptron',
-    # recurrent_neural_network
+    # rnn / gru / lstm
     'GatedRecurrentUnit',
     'LongShortTermMemory',
     'RecurrentNeuralNetwork',
